Altered loan calculations to support "Bitcoin Lightning Mortgages"
Loans are now solely used to purchase more BTC and add to the Stable Receiver allocation with a different fixed rate. This increases Bitcoin Growth and Transaction exposure on the Lightning Network instead of for loaning fiat to the user.

Loaning fiat as a fixed term or flexible credit line using the deposited Bitcoin as collateral can still be a feature added later but that is a completely separate feature.